# openapi.yml
openapi: 3.0.0
info:
  title: ASAP API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/users/me:
    patch:
      summary: Update current user
      operationId: updateCurrentUser
      tags:
        - users
      security:
        - bearerAuth: []
      description: Update current user's profile information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      summary: Get current user
      operationId: getCurrentUser
      tags:
        - users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/users/me/password:
    put:
      summary: Update current user's password
      operationId: updateCurrentUserPassword
      tags:
        - users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordRequest"
      responses:
        "200":
          description: Password updated successfully
        "400":
          description: Invalid request or incorrect current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/users/me/auth-methods:
    get:
      summary: Get current user's authentication methods configuration
      operationId: getCurrentUserAuthMethods
      tags:
        - users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user's authentication methods
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAuthMethods"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update current user's authentication methods
      operationId: updateCurrentUserAuthMethods
      tags:
        - users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAuthMethodsRequest"
      responses:
        "200":
          description: Updated authentication methods
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAuthMethods"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/users/me/email-preferences:
    get:
      summary: Get current user's email preferences
      operationId: getCurrentUserEmailPreferences
      tags:
        - users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user's email preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEmailPreferences"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update current user's email preferences
      operationId: updateCurrentUserEmailPreferences
      tags:
        - users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Updated email preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEmailPreferences"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/public/users/email-preferences:
    put:
      summary: Public update of user's email preferences using token
      operationId: updatePublicUserEmailPreferences
      tags:
        - users
      description: Update a user's email preferences using a token provided in emails. Token validity is signature-based; expiration is not enforced.
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: Signed token used to authorize updates to the user's preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Updated email preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEmailPreferences"
        "400":
          description: Invalid request or token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: Public get of user's email preferences using token
      operationId: getPublicUserEmailPreferences
      tags:
        - users
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: Signed token used to authorize reading the user's preferences
      responses:
        "200":
          description: Current user's email preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEmailPreferences"
        "400":
          description: Invalid request or token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/users:
    post:
      summary: Create user
      operationId: createAdminUser
      tags:
        - users
      security:
        - bearerAuth: []
      description: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "200":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      summary: Find all users
      operationId: findAdminUsers
      tags:
        - users
      security:
        - bearerAuth: []
      description: Find all users
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
          description: Search query
        - name: organizationId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Organization ID filter
        - name: organizationName
          in: query
          required: false
          schema:
            type: string
          description: Organization name filter
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
          description: Page number (starts from 0)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
          description: Number of items per page
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
          description: Field to sort by
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
          description: Sort order
      responses:
        "200":
          description: Users response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUserResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/users/{id}:
    get:
      summary: Get user by ID
      operationId: getAdminUserById
      tags:
        - users
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: Update user
      operationId: updateAdminUser
      tags:
        - users
      security:
        - bearerAuth: []
      description: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete user by ID
      operationId: deleteAdminUserById
      tags:
        - users
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
      responses:
        "200":
          description: User deleted successfully
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/users:
    post:
      summary: Create user
      operationId: createUser
      tags:
        - users
      security:
        - bearerAuth: []
      description: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "200":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      summary: Find all users
      operationId: findUsers
      tags:
        - users
      security:
        - bearerAuth: []
      description: Find all users
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
          description: Search query
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
          description: Page number (starts from 0)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
          description: Number of items per page
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
          description: Field to sort by
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
          description: Sort order
      responses:
        "200":
          description: Users response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUserResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/users/{id}:
    get:
      summary: Get user by ID
      operationId: getUserById
      tags:
        - users
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: Update user
      operationId: updateUser
      tags:
        - users
      security:
        - bearerAuth: []
      description: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete user by ID
      operationId: deleteUserById
      tags:
        - users
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
      responses:
        "200":
          description: User deleted successfully
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ========= Organizations =========
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: integer
          description: Error code
        message:
          type: string
          description: Human-readable error message
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred
        path:
          type: string
          description: Request path that caused the error
    Timestamp:
      type: string
      format: date-time
      description: Timestamp in ISO 8601 format
    PaginatedResponse:
      type: object
      properties:
        page:
          type: integer
          description: Current page number (zero-based)
        size:
          type: integer
          description: Number of items per page
        totalElements:
          type: integer
          format: int64
          description: Total number of elements
        totalPages:
          type: integer
          description: Total number of pages

    # ========= User Schemas =========
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the user
        email:
          type: string
          description: Email address of the user
        firstName:
          type: string
          description: User's first name
        lastName:
          type: string
          description: User's last name
        phone:
          type: string
          description: User's phone number
        avatarUrl:
          type: string
          format: uri
          description: URL to the user's avatar image
        roles:
          type: array
          items:
            $ref: "#/components/schemas/UserRole"
          description: User's roles
        isActive:
          type: boolean
          description: Whether the user is active
        twoFactorAuthEnabled:
          type: boolean
          description: Whether two-factor authentication is enabled
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
    CurrentUser:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            organizationId:
              type: string
              format: uuid
              description: The current organization ID the user is operating in
    UserRole:
      type: string
      enum:
        - PLATFORM_ADMIN
        - OWNER
        - ADMIN
        - MEMBER
        - VIEWER
        - ACCOUNTANT
    CreateUserRequest:
      type: object
      required:
        - email
        - firstName
        - roles
      properties:
        organizationId:
          type: string
          format: uuid
          description: Organization ID
        email:
          type: string
          format: email
          description: Email address of the user
        firstName:
          type: string
          description: User's first name
        lastName:
          type: string
          description: User's last name
        phone:
          type: string
          description: User's phone number
        roles:
          type: array
          items:
            $ref: "#/components/schemas/UserRole"
        sendEmail:
          type: boolean
          description: Whether to send an invitation email to the user
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
          nullable: true
          description: User's first name
        lastName:
          type: string
          nullable: true
          description: User's last name
        phone:
          type: string
          nullable: true
          description: User's phone number
        avatarUrl:
          type: string
          format: uri
          nullable: true
          description: URL to the user's avatar image
        roles:
          nullable: true
          type: array
          items:
            $ref: "#/components/schemas/UserRole"
          description: User's roles
      description: Request to update user profile information
    UpdatePasswordRequest:
      type: object
      properties:
        oldPassword:
          type: string
          description: Current password
        newPassword:
          type: string
          description: New password
        twoFactorAuthCode:
          type: string
          description: 2FA code (if applicable)
    UserAuthMethods:
      type: object
      properties:
        passwordEnabled:
          type: boolean
          description: Whether password authentication is enabled
        loginLinkEnabled:
          type: boolean
          description: Whether login link authentication is enabled
        googleEnabled:
          type: boolean
          description: Whether Google OAuth authentication is enabled
    UpdateAuthMethodsRequest:
      type: object
      properties:
        passwordEnabled:
          type: boolean
          description: Enable or disable password authentication
        loginLinkEnabled:
          type: boolean
          description: Enable or disable login link authentication
        googleEnabled:
          type: boolean
          description: Enable or disable Google OAuth authentication
        twoFactorAuthCode:
          type: string
          description: 2FA code (required if 2FA is active)
    UserEmailPreferences:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User's email address
        preferences:
          type: object
          additionalProperties: true
    PaginatedUserResponse:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/User"
              description: List of users

    # ========= Organization Schemas =========
...